requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pandas==2.1.2
numpy==1.24.3
python-dotenv==1.0.0

# Database and vector store
neo4j==5.14.0
chromadb==0.4.15

# LLM and embeddings
openai==1.3.0

# Chemical data processing
rdkit==2023.9.1

# Graph visualization
networkx==3.2.1
pyvis==0.3.2
plotly==5.18.0

# Additional utilities
requests==2.31.0
python-multipart==0.0.6